+++ sdk-extensions/autoconfigure/src/main/java/io/opentelemetry/sdk/autoconfigure/OpenTelemetryResourceAutoConfiguration.java
@@ -8,11 +8,18 @@ package io.opentelemetry.sdk.autoconfigure;
 import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
 import io.opentelemetry.sdk.autoconfigure.spi.ResourceProvider;
 import io.opentelemetry.sdk.resources.Resource;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.ServiceLoader;
 import java.util.Set;
+import java.util.function.BiFunction;
 
-/** Auto-configuration for the OpenTelemetry {@link Resource}. */
+/**
+ * Auto-configuration for the OpenTelemetry {@link Resource}.
+ *
+ * @deprecated Use {@link AutoConfiguredOpenTelemetrySdk#getResource()}.
+ */
+@Deprecated
 public final class OpenTelemetryResourceAutoConfiguration {
 
   /**
@@ -20,16 +27,29 @@ public final class OpenTelemetryResourceAutoConfiguration {
    *
    * <p>This method will auto-configure the returned {@link Resource} using system properties and
    * environment variables.
+   *
+   * @deprecated Use {@link AutoConfiguredOpenTelemetrySdk#getResource()}.
    */
+  @Deprecated
   public static Resource configureResource() {
-    return configureResource(DefaultConfigProperties.get());
+    return configureResource(DefaultConfigProperties.get(Collections.emptyMap()));
   }
 
   /**
    * Returns a {@link Resource} automatically initialized through recognized system properties and
    * environment variables.
+   *
+   * @deprecated Use {@link AutoConfiguredOpenTelemetrySdk#getResource()}.
    */
+  @Deprecated
   public static Resource configureResource(ConfigProperties config) {
+    return configureResource(config, (a, unused) -> a);
+  }
+
+  @SuppressWarnings("deprecation") // Uses class which will be made package private
+  static Resource configureResource(
+      ConfigProperties config,
+      BiFunction<? super Resource, ConfigProperties, ? extends Resource> resourceCustomizer) {
     Resource result = Resource.getDefault();
 
     // TODO(anuraaga): We use a hyphen only once in this artifact, for
@@ -46,7 +66,7 @@ public final class OpenTelemetryResourceAutoConfiguration {
 
     result = result.merge(EnvironmentResource.create(config));
 
-    return result;
+    return resourceCustomizer.apply(result, config);
   }
 
   private OpenTelemetryResourceAutoConfiguration() {}