+++ sdk-extensions/autoconfigure/src/test/java/io/opentelemetry/sdk/autoconfigure/PropagatorConfigurationTest.java
@@ -17,7 +17,7 @@ class PropagatorConfigurationTest {
   void defaultPropagators() {
     ContextPropagators contextPropagators =
         PropagatorConfiguration.configurePropagators(
-            DefaultConfigProperties.createForTest(Collections.emptyMap()));
+            DefaultConfigProperties.createForTest(Collections.emptyMap()), (a, unused) -> a);
 
     assertThat(contextPropagators.getTextMapPropagator().fields())
         .containsExactlyInAnyOrder("traceparent", "tracestate", "baggage");