+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/state/SynchronousMetricStorage.java
@@ -46,9 +46,6 @@ public interface SynchronousMetricStorage extends MetricStorage, WriteableMetric
       return empty();
     }
     return new DefaultSynchronousMetricStorage<>(
-        metricDescriptor,
-        aggregator,
-        view.getAttributesProcessor(),
-        view.getAggregation().getConfiguredTemporality());
+        metricDescriptor, aggregator, view.getAttributesProcessor());
   }
 }